requirements.txt
# API and Data Processing
requests==2.31.0
pandas==2.0.3
alpaca-trade-api==3.0.0
python-dotenv==1.0.0
numpy==1.25.2
yfinance==0.2.28
praw==7.7.1
aiohttp==3.8.5

# Technical Analysis
ta==0.10.2
TA-Lib==0.4.28

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.2

# Visualization and Analysis
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.66.1

# Infrastructure
schedule==1.2.0
psutil==5.9.5

# Testing
pytest==7.4.0
unittest-mock==1.3.0

# Deployment
gunicorn==21.2.0
flask==2.3.3
